requirements.txt
# TicoViz Corporation v2 - Dependencias (Railway)
# pip install -r requirements.txt

# Telegram Bot
python-telegram-bot==21.3

# Web Server
Flask==3.1.0
gunicorn==23.0.0

# Web Scraping (MP4)
beautifulsoup4==4.12.3
requests==2.32.3
lxml==5.3.0

# Generacion de documentos
reportlab==4.2.5
python-pptx==1.0.2
openpyxl==3.1.5
Pillow==11.0.0

# HTTP async (para Groq + Gemini API)
aiohttp==3.11.11

# Utilidades
python-dotenv==1.0.1
